Registration added

+++ src/main/java/com/example/petproject/repository/PyramidBilliardCuesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface PyramidBilliardCuesRepository extends CrudRepository<PyramidBilliardCues, Long> {
+public interface PyramidBilliardCuesRepository extends CrudRepository<PyramidBilliardCues, Integer> {
 
     List<PyramidBilliardCues> findByProducer(String producer);
 }
